Extract PropertyBagToTypeInstanceFactory
